Dimensionally-Reduction/Principal-Component-Analysis(PCA)/requirements.txt
numpy
pandas
statsmodel
os
PIL
sklearn
plotly